fix(wip): move JSX DOM types back to `@vue/runtime-dom`
Drops `vue/jsx-runtime/dom`.
The previous commits didn't work because imported (and re-exported)
types cannot be augmented.
So I give up on the idea of moving the DOM types to `vue/jsx-runtime`.
They don't pollute the global namespace anyway.

There's still one caveat that I have to import the
`IntrinsicElementAttributes` type from
`@vue/runtime-dom/dist/runtime-dom` rather than using the package name
as entry. Seems because in this monorepo, TypeScript always picks up
the source file (`index.ts`) over the built file
(`dist/runtime-dom.d.ts` as specified in the `types` field in
`package.json`)
