+++ packages/runtime-dom/package.json
@@ -8,7 +8,6 @@
   "unpkg": "dist/runtime-dom.global.js",
   "files": [
     "index.js",
-    "jsx.d.ts",
     "dist"
   ],
   "sideEffects": false,
