--- packages/vue/jsx-runtime/dom.d.ts
@@ -1,1 +1,0 @@
-export * from '@vue/runtime-dom/jsx'
