+++ packages/vue/jsx.d.ts
@@ -1,9 +1,10 @@
 // global JSX namespace registration
 // somehow we have to copy=pase the jsx-runtime types here to make TypeScript happy
-import { VNode, VNodeRef } from '@vue/runtime-dom'
-import { IntrinsicElementAttributes } from './jsx-runtime/dom'
-
-export * from './jsx-runtime/dom'
+import type {
+  VNode,
+  VNodeRef,
+  IntrinsicElementAttributes
+} from '@vue/runtime-dom/dist/runtime-dom'
 
 export type ReservedProps = {
   key?: string | number | symbol
